+++ packages/gatsby-recipes/src/resources.js
@@ -4,7 +4,7 @@ export * as GatsbyPage from "./providers/gatsby/page"
 export * as GatsbyPlugin from "./providers/gatsby/plugin"
 export * as GatsbyShadowFile from "./providers/gatsby/shadow-file"
 export * as GatsbySiteMetadata from "./providers/gatsby/site-metadata"
-export * as GatsbyExperiments from "./providers/gatsby/experiments"
+export * as GatsbyExperiment from "./providers/gatsby/experiment"
 export * as NPMPackage from "./providers/npm/package"
 export * as NPMScript from "./providers/npm/script"
 export * as NPMPackageJson from "./providers/npm/package-json"
